requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.6.1
pydantic-settings>=2.4.0
redis==5.0.1
httpx==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
google-generativeai==0.3.0
langdetect==1.0.9
pytest>=8.2
pytest-asyncio>=1.2.0